requirements.txt
Django==3.1
django-cors-headers==3.6.0
djangorestframework==3.12.2
django_heroku
gunicorn